@@ -17,21 +17,32 @@
     "require": {
         "php": "^7.1",
         "ext-tokenizer": "*",
-        "nikic/php-parser": "^3.1",
-        "padraic/phar-updater": "^1.0",
-        "symfony/console": "^4.0",
-        "symfony/filesystem": "^4.0",
-        "symfony/finder": "^4.0",
-        "symfony/process": "^4.0",
-        "twig/twig": "^2.4"
+        "nikic/php-parser": "^4.10.5",
+        "padraic/phar-updater": "^1.0.6",
+        "symfony/console": "^4.4.24",
+        "symfony/filesystem": "^4.4.22",
+        "symfony/finder": "^4.4.24",
+        "symfony/process": "^4.4.22",
+        "twig/twig": "^2.13.1",
+        "webmozart/assert": "^1.9.1"
     },
     "require-dev": {
         "ext-imagick": "*",
-        "lupka/phpunit-compare-images": "^1.0",
-        "phpunit/phpunit": "^6.5"
+        "ergebnis/composer-normalize": "^2.8.0",
+        "friendsofphp/php-cs-fixer": "^3.0.0",
+        "lupka/phpunit-compare-images": "^1.0.0",
+        "maglnet/composer-require-checker": "^2.0.0",
+        "phpro/grumphp": "^0.16.2",
+        "phpstan/extension-installer": "^1.1.0",
+        "phpstan/phpstan": "^0.12.88",
+        "phpstan/phpstan-strict-rules": "^0.12.9",
+        "phpunit/phpunit": "^7.5.20"
     },
     "config": {
         "optimize-autoloader": true,
+        "platform": {
+            "php": "7.1.30"
+        },
         "sort-packages": true
     },
     "extra": {
@@ -46,7 +57,11 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "PhUml\\": "tests"
+            "PhUml\\": [
+                "tests/src",
+                "tests/integration",
+                "tests/unit"
+            ]
         }
     },
     "prefer-stable": true,
