@@ -32,7 +32,7 @@ class MethodDocBlock extends DocBlock
     public function returnType(): TypeDeclaration
     {
         $type = null;
-        if (preg_match(self::$returnExpression, $this->comment, $matches)) {
+        if (preg_match(self::$returnExpression, (string)$this->comment, $matches) === 1) {
             $type = trim($matches[1]);
         }
         return TypeDeclaration::from($type);
@@ -51,7 +51,7 @@ class MethodDocBlock extends DocBlock
 
     private function setParameters(): void
     {
-        if (!preg_match_all(self::$parameterExpression, $this->comment, $matches)) {
+        if (preg_match_all(self::$parameterExpression, (string)$this->comment, $matches) < 1) {
             return;
         }
         foreach ($matches[0] as $typeHint) {
@@ -61,7 +61,7 @@ class MethodDocBlock extends DocBlock
 
     private function extractDeclarationFrom(string $typeHint): void
     {
-        if (preg_match(self::$parameterExpression, $typeHint, $match)) {
+        if (preg_match(self::$parameterExpression, $typeHint, $match) === 1) {
             [$_, $type, $parameterName] = $match;
             $this->parameters[$parameterName] = TypeDeclaration::from($type);
         }
