@@ -7,9 +7,9 @@
 
 namespace PhUml\Parser\Code\Builders\Members;
 
+use PhpParser\Node;
 use PhpParser\Node\Const_;
 use PhpParser\Node\Expr\ConstFetch;
-use PhpParser\Node\Scalar;
 use PhpParser\Node\Stmt\ClassConst;
 use PhUml\Code\Attributes\Constant;
 use PhUml\Code\Variables\TypeDeclaration;
@@ -19,6 +19,7 @@ use PhUml\Code\Variables\TypeDeclaration;
  */
 class ConstantsBuilder
 {
+    /** @var string[] */
     private static $types = [
         'integer' => 'int',
         'double' => 'float',
@@ -26,15 +27,15 @@ class ConstantsBuilder
     ];
 
     /**
-     * @param \PhpParser\Node[] $classAttributes
+     * @param Node[] $classAttributes
      * @return Constant[]
      */
     public function build(array $classAttributes): array
     {
-        $constants = array_filter($classAttributes, function ($attribute) {
+        $constants = array_filter($classAttributes, static function ($attribute): bool {
             return $attribute instanceof ClassConst;
         });
-        return array_map(function (ClassConst $constant) {
+        return array_map(function (ClassConst $constant): Constant {
             return new Constant(
                 $constant->consts[0]->name,
                 TypeDeclaration::from($this->determineType($constant->consts[0]))
@@ -44,7 +45,7 @@ class ConstantsBuilder
 
     private function determineType(Const_ $constant): ?string
     {
-        if ($constant->value instanceof Scalar) {
+        if (property_exists($constant->value, 'value')) {
             return self::$types[\gettype($constant->value->value)];
         }
         if ($constant->value instanceof ConstFetch
