@@ -32,8 +32,9 @@ class ClassVisitor extends NodeVisitorAbstract
 
     public function leaveNode(Node $node)
     {
-        if ($node instanceof Class_) {
+        if ($node instanceof Class_ && !$node->isAnonymous()) {
             $this->codebase->add($this->builder->build($node));
         }
+        return null;
     }
 }
